Improve mismatch check between profile name and export method
Assume that profile names are generated by match and follow the same pattern
of having the string "match" at the start.

Avoid searching for export method anywhere in the profile name, but
instead assumes that the profile names follow the same pattern of
having the string "match" at the start, which follows the naming
convention of match itself.